refactor modals
