+++ public/css/history.css
@@ -1,0 +1,92 @@
+#history-wrap {
+  overflow: auto;
+  height: 500px;
+  letter-spacing: -0.3px;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+  color: #fff;
+}
+#history-wrap .history-item {
+  margin-bottom: 35px;
+  border-bottom: 1px solid #8ffdfd;
+}
+#history-wrap .history-item .item-section {
+  padding: 0;
+  text-align: center;
+}
+#history-wrap .history-item p {
+  font-size: 17px;
+  margin-bottom: 8px;
+  text-align: center;
+}
+#history-wrap .history-item p span {
+  position: relative;
+  top: 2px;
+}
+#history-wrap .history-item p .blue-text {
+  padding: 0 !important;
+}
+#history-wrap .history-item .amount {
+  display: inline-block;
+  margin-bottom: 8px;
+}
+#history-wrap .history-item .arrow-img {
+  max-height: 18px;
+  margin-right: 10px;
+}
+#history-wrap .history-item .green-text {
+  color: rgba(14, 200, 193, 1);
+  font-size: 17px;
+}
+#history-wrap .history-item .yellow-text {
+  color: #e9d70b;
+}
+#history-wrap .history-item .tx-id-wrap {
+  display: flex;
+  justify-content: center;
+}
+#history-wrap .history-item .tx-id-wrap span {
+  display: inline-block;
+  text-transform: initial;
+  text-align: left;
+  padding: 0 15px 0 0;
+}
+#history-wrap .history-item .tx-id {
+  background: transparent;
+  border: none;
+  text-transform: initial;
+  font-size: 16px;
+  padding: 0;
+  display: inline-block;
+  letter-spacing: 0;
+  margin-bottom: 0;
+  color: #54bcd9;
+  width: calc(100% - 105px);
+  text-align: left;
+  white-space: nowrap;
+}
+#history-wrap .history-item .tx-id:hover {
+  box-shadow: none !important;
+}
+#history-wrap .question-wrap {
+  position: absolute;
+  top: 15px;
+  left: 15px;
+  width: 32px;
+  height: 32px;
+  border: 1px solid #8ffdfd;
+  border-radius: 50%;
+  text-align: center;
+  transition: all ease 0.3s;
+  -o-transition: all ease 0.3s;
+  -moz-transition: all ease 0.3s;
+  -webkit-transition: all ease 0.3s;
+  -webkit-touch-callout: none; /* iOS Safari */
+  -webkit-user-select: none; /* Safari */
+  -khtml-user-select: none; /* Konqueror HTML */
+  -moz-user-select: none; /* Firefox */
+  -ms-user-select: none; /* Internet Explorer/Edge */
+  user-select: none; /* Non-prefixed version, currently supported by Chrome and Opera */
+}
+#history-wrap .pagination {
+  bottom: 130px;
+}
