+++ public/css/item.css
@@ -33,7 +33,7 @@
 }
 .item-wrap h2 {
   color: rgba(255, 255, 255, 0.95);
-  font-size: 44px;
+  font-size: 30px;
   text-align: center;
   text-transform: uppercase;
   margin: 40px 0 10px;
