+++ public/css/sidebar.css
@@ -1,0 +1,275 @@
+.sidebar {
+  background: rgba(15, 29, 54, 0.9);
+  display: flex;
+  position: fixed;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: -100%;
+  z-index: 5;
+  -webkit-transition: all ease 0.4s;
+  -moz-transition: all ease 0.4s;
+  transition: all ease 0.4s;
+}
+.sidebar h2 {
+  color: #fff;
+  margin: 40px 0 50px;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+  font-size: 22px;
+  text-transform: initial;
+  letter-spacing: 0;
+}
+.sidebar.active {
+  left: 0;
+}
+.sidebar #nav {
+  width: 25%;
+  border-right: 1px solid #54bcd9;
+}
+.sidebar #nav button {
+  text-align: center;
+  background: url(../images/login-bg.png) no-repeat top left / cover;
+  border: 2px solid #54bcd9;
+  padding: 7px 13px 8px 12px;
+  margin-top: 80px;
+  margin-left: 31px;
+  display: flex;
+  width: 48px;
+  justify-content: center;
+}
+.sidebar #nav button img {
+  max-height: 18px;
+}
+.sidebar #nav ul {
+  list-style-type: none;
+  margin: 20px 0 0;
+  padding-left: 0;
+}
+.sidebar #nav ul li {
+  background: rgba(15, 29, 54, 1);
+  border-top: 1px solid #54bcd9;
+  padding: 11px 0 8px 30px;
+  color: #fff;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+  font-size: 17px;
+  -webkit-transition: all ease 0.3s;
+  -moz-transition: all ease 0.3s;
+  transition: all ease 0.3s;
+}
+.sidebar #nav ul li:last-of-type {
+  border-bottom: 1px solid #54bcd9;
+}
+.sidebar #nav ul li:hover {
+  background: rgba(0, 0, 0, 1);
+  cursor: pointer;
+}
+.sidebar #nav ul li.active {
+  background: rgba(0, 0, 0, 1);
+}
+.sidebar #nav ul li img {
+  max-width: 20px;
+  margin-right: 20px;
+  position: relative;
+  top: -1px;
+}
+.sidebar #content {
+  background: rgba(0, 0, 0, 0.4);
+  position: relative;
+  width: 75%;
+  padding: 0 50px;
+  overflow: auto;
+}
+.sidebar #content .label-wrap {
+  text-align: center;
+  position: relative;
+  z-index: 3;
+  margin: 15px 0 10px;
+}
+.sidebar #content .label-wrap label {
+  width: initial !important;
+  text-align: center;
+  font-size: 20px;
+  color: rgba(255, 255, 255, 1);
+  margin-bottom: 0;
+  font-weight: 600;
+  line-height: 1.2;
+  position: relative;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+}
+.sidebar #content label button {
+  border-radius: 4px;
+  background: url(../images/login-bg.png) no-repeat top left / cover;
+  border: 2px solid #54bcd9;
+  color: #fff;
+  font-size: 13px;
+  font-weight: 700;
+  text-transform: uppercase;
+  letter-spacing: 1.26px;
+  line-height: 1.2;
+  position: absolute;
+  right: 0;
+  top: -3px;
+  padding: 4px 10px;
+  -webkit-transition: all 0.2s ease-in;
+  -moz-transition: all 0.2s ease-in;
+  transition: all 0.2s ease-in;
+}
+.sidebar #content .label-wrap .question-wrap {
+  position: absolute;
+  top: -2px;
+  left: 0;
+  width: 32px;
+  height: 32px;
+  border: 1px solid #8ffdfd;
+  border-radius: 50%;
+  text-align: center;
+  z-index: 4;
+  transition: all ease 0.3s;
+  -o-transition: all ease 0.3s;
+  -moz-transition: all ease 0.3s;
+  -webkit-transition: all 0.3s ease;
+}
+.sidebar #content .label-wrap .question-wrap:hover {
+  -webkit-box-shadow: 0 0 20px 0 rgba(0, 199, 236, 0.6);
+  box-shadow: 0 0 20px 0 rgba(0, 199, 236, 0.6);
+  cursor: pointer;
+}
+.sidebar #content .label-wrap .question-wrap span {
+  color: #fff;
+  position: relative;
+  top: -7px;
+  left: -0.5px;
+  font-size: 19px;
+  transition: all ease 0.3s;
+  -o-transition: all ease 0.3s;
+  -moz-transition: all ease 0.3s;
+  -webkit-transition: all ease 0.3s;
+}
+.sidebar #content .label-wrap .copy-btn {
+  background: url(../images/login-bg.png) no-repeat top left / cover;
+  padding: 5px 9px 4px;
+  line-height: 1.2;
+  text-transform: uppercase;
+  position: absolute;
+  top: 0px;
+  right: 0;
+  z-index: 4;
+  font-size: 12px;
+  border: 2px solid #54bcd9;
+  border-radius: 7px;
+  color: #fff;
+  font-weight: 600;
+  letter-spacing: 0.3px;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+  -webkit-transition: all 0.2s ease-in;
+  -moz-transition: all 0.2s ease-in;
+  transition: all 0.2s ease-in;
+}
+.sidebar #content textarea {
+  width: 100%;
+  background: transparent;
+  border: 0px none;
+  border-radius: 0px;
+  color: rgba(255, 255, 255, 0.8);
+  font-size: 16px;
+  padding: 0 0 5px !important;
+  border-bottom: 1px solid #8ffdfd;
+  margin: 15px 0;
+  resize: none;
+  text-align: center;
+  color: #fff;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+}
+.sidebar #content input {
+  display: inline-block;
+  border: 0px none;
+  padding: 5px 10px 7px;
+  margin: 0 0 15px !important;
+  font-size: 16px;
+  background: transparent;
+  border-bottom: 1px solid #8ffdfd;
+  border-radius: 0;
+  color: #fff !important;
+  width: 100%;
+  max-height: 90px;
+  text-align: center;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+  transition: all ease 0.3s;
+  -o-transition: all ease 0.3s;
+  -moz-transition: all ease 0.3s;
+  -webkit-transition: all ease 0.3s;
+}
+.sidebar #content textarea::-webkit-input-placeholder,
+.sidebar #content input::-webkit-input-placeholder {
+  /* Chrome/Opera/Safari */
+  color: rgba(255, 255, 255, 0.9);
+  font-size: 18px !important;
+  letter-spacing: 0.3px;
+  text-align: left;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+  line-height: 1.2;
+  word-break: break-all;
+}
+.sidebar #content textarea::-moz-placeholder,
+.sidebar #content input::-moz-placeholder {
+  /* Firefox 19+ */
+  color: rgba(255, 255, 255, 0.9);
+  font-size: 18px !important;
+  letter-spacing: 0.3px;
+  text-align: left;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+}
+.sidebar #content textarea:-moz-placeholder,
+.sidebar #content input:-moz-placeholder {
+  /* Firefox 18- */
+  color: rgba(255, 255, 255, 0.9);
+  font-size: 18px !important;
+  letter-spacing: 0.3px;
+  text-align: left;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+}
+.sidebar #content textarea:-ms-input-placeholder,
+.sidebar #content input:-ms-input-placeholder {
+  /* IE 10+ */
+  color: rgba(255, 255, 255, 0.9);
+  font-size: 18px !important;
+  letter-spacing: 0.3px;
+  text-align: left;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+}
+.sidebar .label-wrap {
+  text-align: center;
+  position: relative;
+  margin-bottom: 3px;
+}
+.sidebar label {
+  width: 100%;
+  text-transform: capitalize;
+  font-size: 18px;
+  font-weight: 400;
+  margin: 10px 0 0;
+  position: relative;
+  z-index: 2;
+  color: #fff;
+  text-align: center;
+  font-weight: 600;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
+}
+.sidebar label .question-wrap {
+  position: absolute;
+  top: -10px;
+  right: 0;
+  width: 32px;
+  height: 32px;
+  border: 1px solid #8ffdfd;
+  border-radius: 50%;
+  transition: all ease 0.3s;
+  -o-transition: all ease 0.3s;
+  -moz-transition: all ease 0.3s;
+  -webkit-transition: all ease 0.3s;
+}
+.sidebar label .question-wrap:hover {
+  -webkit-box-shadow: 0 0 20px 0 rgba(0, 199, 236, 0.6);
+  box-shadow: 0 0 20px 0 rgba(0, 199, 236, 0.6);
+  cursor: pointer;
+}
