+++ public/index.html
@@ -13,6 +13,10 @@
     <link rel="stylesheet" href="css/item.css" />
     <link rel="stylesheet" href="css/wallet.css" />
     <link rel="stylesheet" href="css/header.css" />
+    <link rel="stylesheet" href="css/sidebar.css" />
+    <link rel="stylesheet" href="css/contacts.css" />
+    <link rel="stylesheet" href="css/history.css" />
+    <link rel="stylesheet" href="css/seed&keys.css" />
     <link rel="stylesheet" href="css/mobile.css" />
     <!--
       Notice the use of %PUBLIC_URL% in the tag above.
